Unresolvable addresses return an empty list.
Don't want exceptions propagating out of here. We can handle the error
case higher up. This is similar to the case where, for example, we
request a dns6 resolution on a name which only has an ip4 address. If
people need detailed reporting on DNS failures they can investigate
afterwards