Hook up Chat protocol
Chat is a bidirectional protocol for sending point-to-point messages.
Use discovery to find out peers. When new peers are found, use Identify to see if
they support the Chat protocol. If they do, then connect to them (or use the connection
the peer has just opened to us). Messages entered on the console are sent to all
connected chat peers.