Consolidate both Gradle projects into one
Combine the "example" projects into the main project. I needed to move
the main library into it's own module so the other libraries could
exist. This allows a developer to open this project in IntelliJ and
build all of the projects easily.

There's still more clean-up to do such as consolidating the Kotlin
versions to one top level declaration, but I'm trying to make the
minimal amount of changes to get this to work.

Closes https://github.com/libp2p/jvm-libp2p/issues/153
Closes https://github.com/libp2p/jvm-libp2p/issues/152