Suppress unchecked casts in two places
These were the warnings:

```
P2PChannelHandler.kt: (29, 18): Unchecked cast: P2PChannel to TChannel
HostImpl.kt: (98, 83): Unchecked cast: ProtocolBinding<Any>? to ProtocolBinding<TController>
```

From what I've learned, it is difficult for Kotlin to cast generics. For the
first warning, I tried to rewrite the cast so that it would be safer but I was
unsuccessful so I just added a suppression decorator. For the second warning, I
believe the decorator was just on the wrong line. If removed it wouldn't
generate an additional error.